pyproject.toml
[project]
name = "workweek_survey"
version = "0.1.0"
description = "A fast survey app for cataloging team members’ weekly tasks."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "pyyaml",
  "jinja2",
  "pydantic-settings",
]
[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
  "pytest-asyncio",
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
